requirements.txt
# ------------------------
# 核心必装库（项目骨架运行必需）
# ------------------------
numpy>=1.24                 # 数值计算、坐标变换
open3d>=0.17                # 点云/mesh操作，PLY读写
pillow>=9.0                 # 图像保存/处理（RGB / depth 占位）

# ------------------------
# 可选库（扩展功能、复杂 mesh 或渲染）
# ------------------------
trimesh>=3.21               # 高级 mesh 操作，可选
PyYAML>=6.0                  # yaml 配置文件解析（如果 dataset.yaml / lidar.yaml 需要解析）
matplotlib>=3.7              # 可视化（调试点云 / shadow mask）
scipy>=1.11                  # 高级数学 / 坐标变换可选
